+++ src/index.ts
@@ -7,8 +7,8 @@ export * from './legacy/compiler'
 export * from './legacy/ts-jest-transformer'
 export * from './legacy/config/config-set'
 export * from './presets/create-jest-preset'
-export { nodeModulesTransformPattern } from './presets/node-modules-transform-pattern'
-export type { NodeModulesTransformPatternOptions } from './presets/node-modules-transform-pattern'
+export { nodeModulesTransformPattern } from './patterns/node-modules-transform-pattern'
+export type { NodeModulesTransformPatternOptions } from './patterns/node-modules-transform-pattern'
 export * from './raw-compiler-options'
 export * from './utils'
 export * from './types'
