+++ src/patterns/node-modules-transform-pattern.ts
